Drop Elasticsearch5Index and make Elasticsearch6Index the base implementation
